FoodCardViewComponents 제거
- 제거에 따른 변경 사항 반영
